fix(cmake): Fix GCC build on Intel macOS (build failure, warnings, and small-angle std::cos(float) issue) (#591)
* [Codex] Fix build failure on GCC on intel macOS

* [Codex] Fixed macOS on intel GCC warnings

It found the following bits of related technical debt to review later:
  * include/sw/universal/number/unum2/unum2_impl.hpp:412: mixed-domain shift_count
  * include/sw/universal/number/unum2/lattice.hpp:122: suspicious pre-existing _N >> 1 vs _N_half / _N_half logic

* [Codex] Generalized std::cos(float) to fix rounding of small inputs to zero on intel.

* [ChatGPT] better comments and variable names

---------

+++ include/sw/universal/number/unum2/op_matrix.hpp
@@ -23,13 +23,13 @@ private:
 
 public:
     op_matrix(size_t N) : _N { N } {
-        for(int i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
+        for(size_t i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
             _matrices[i] = nullptr;
             _present_flags[i] = nullptr;
         }
 
         try {
-            for(int i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
+            for(size_t i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
                 _matrices[i] = new unum2<T>[N * N];
                 _present_flags[i] = new bool[N * N];
             }
@@ -76,7 +76,7 @@ public:
 
 private:
     void cleanup() {
-        for(int i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
+        for(size_t i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
             delete[] _matrices[i];
             delete[] _present_flags[i];
